Move plugins to central place tests/common/plugins
Move plugins required by tests/conftest.py to a central place tests/common/plugins:
* ansible_fixture.py
* fib.py
* loganalyzer
* ptfadapter
* dut_monitor
