+++ tests/common/plugins/dut_monitor/README.md
