+++ tests/common/plugins/dut_monitor/thresholds.yml
