+++ tests/common/plugins/loganalyzer/loganalyzer_common_ignore.txt
@@ -1,0 +1,1 @@
+../../../../ansible/roles/test/files/tools/loganalyzer/loganalyzer_common_ignore.txt
