+++ tests/common/plugins/loganalyzer/loganalyzer_common_match.txt
@@ -1,0 +1,1 @@
+../../../../ansible/roles/test/files/tools/loganalyzer/loganalyzer_common_match.txt
