+++ tests/common/plugins/loganalyzer/system_msg_handler.py
@@ -1,0 +1,1 @@
+../../../../ansible/roles/test/files/tools/loganalyzer/loganalyzer.py
