+++ tests/common/plugins/ptfadapter/templates/ptf_nn_agent.conf.ptf.j2
