--- tests/loganalyzer/__init__.py
@@ -1,1 +1,0 @@
-from .loganalyzer import LogAnalyzer, COMMON_MATCH, COMMON_IGNORE, COMMON_EXPECT, LogAnalyzerError
