--- tests/loganalyzer/loganalyzer_common_ignore.txt
@@ -1,1 +1,0 @@
-../../ansible/roles/test/files/tools/loganalyzer/loganalyzer_common_ignore.txt
