--- tests/loganalyzer/loganalyzer_common_match.txt
@@ -1,1 +1,0 @@
-../../ansible/roles/test/files/tools/loganalyzer/loganalyzer_common_match.txt
