--- tests/loganalyzer/system_msg_handler.py
@@ -1,1 +1,0 @@
-../../ansible/roles/test/files/tools/loganalyzer/loganalyzer.py
