--- tests/ptfadapter/__init__.py
@@ -1,3 +1,0 @@
-from ptfadapter import PtfTestAdapter
-
-__all__ = ['PtfTestAdapter']
